swagger-next-doc works just fine
