+++ package.json
@@ -37,6 +37,7 @@
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "sonner": "^2.0.3",
+    "swagger-ui-react": "^5.25.2",
     "tailwind-merge": "^3.2.0",
     "zustand": "^5.0.3"
   },
