+++ src/app/api/users/route.ts
@@ -1,5 +1,15 @@
 import prisma from '@/lib/server/db/prisma';
 
+/**
+ * @openapi
+ * /api/users:
+ *   get:
+ *     description: Returns the hello world
+ *     responses:
+ *       200:
+ *         description: Hello World!
+ */
+
 export async function GET() {
     try {
         const users = await prisma.user.findMany();
