+++ src/app/docs/page.tsx
@@ -1,8 +1,11 @@
-'use client';
-import SwaggerUI from 'swagger-ui-react';
-import 'swagger-ui-react/swagger-ui.css';
-import swaggerSpec from '../../../swagger';
+import { getApiDocs } from "@/lib/swagger";
+import ReactSwagger from "./react-swagger";
 
-export default function Docs() {
-  return <SwaggerUI spec={swaggerSpec} />;
+export default async function IndexPage() {
+  const spec = await getApiDocs();
+  return (
+    <section className="container">
+      <ReactSwagger spec={spec} />
+    </section>
+  );
 }
