--- swagger.ts
@@ -1,16 +1,0 @@
-import { createSwaggerSpec } from 'next-swagger-doc';
-
-export const getApiDocs = async () => {
-  const spec = createSwaggerSpec({
-    apiFolder: "./src/app/api",
-    definition: {
-      openapi: "3.0.0",
-      info: {
-        title: "Next Swagger API Example",
-        version: "1.0",
-      },
-      security: [],
-    },
-  });
-  return spec;
-};
